+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2024-01-09
+
+### Added
+
+- Migrate checkbox to UI
+- Added disabled mode
+- Added stories to storybook
+- Migrate and update tests
+
 ## [0.2.2] -2024-01-09
 
 ### chore
@@ -10,7 +19,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.2] -2023-12-21
 
-### added
+### Added
 
 - Migrate Button to UI
 - Added Button stories to storybook